fix(http): httpStatus should be uint16_t

+++ src/WalterModem.h
@@ -2327,7 +2327,7 @@ typedef struct {
   /*
    * @brief The HTTP status code including our own code to indicate errors during httpDidRing.
    */
-  uint8_t httpStatus;
+  uint16_t httpStatus;
 
   /**
    * @brief The HTTP content length.
